pyproject.toml
[project]
name = "graphrag-vs-vectorrag"
version = "0.1.0"
description = "GraphRAG vs VectorRAG — when knowledge graphs actually help"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "neo4j>=5.25",
    "chromadb>=0.5",
    "numpy>=1.26",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "pyyaml>=6.0",
    "networkx>=3.3",
    "streamlit>=1.38",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "ty>=0.0.1a7",
    "bandit>=1.7",
    "jupyter>=1.1",
    "ipykernel>=6.29",
    "nbconvert>=7.16",
    "nbclient>=0.10",
]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "N", "B", "A", "SIM", "PTH"]
ignore = ["E501", "A003"]

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
